new icons for Solus, Neon, Parabola
added icons for Solus, Neon, Parabola and for the Solus bootloader
(goofiboot, a fork of gummiboot) that has the same icon as Solus.

The reason I use the same icon for Solus and the goofiboot is that
rEFInd is unable to start Solus kernel without manual user
intervention and regular mainteneance on the config file on each
kernel update. I though it was more convenient to have rEFInd boot
goofiboot instead while hiding the folder with Solus kernels.
Since only Solus uses goofiboot and it is their own fork,
I think it's ok if it gets the same icon as Solus.

Parabola and Neon were "targets of opportunity" so to speak.
I saw that people asked for them and found pre-made svg icon
files from their main site or on wikipedia that only needed
resizing, so I thought I could add them too.

I made minor fixes to the /src/render_bitmap.sh because
it was not running on my OpenSUSE Leap 42.2.

I added systemd-boot to the list of stuff that should receive
a "unknown os" icon in the /src/render_bitmap.sh script.
Systemd-boot is a fork of gummiboot that is now part of the
Systemd project, gummiboot has been discontinued.
(I did not delete gummiboot from the list)

While I was at it, I documented the procedure to add new icons in the README.