added better svg files for the two icons from the last commit, still from d4k0 as my edit to the other ones wasn't actually as good as I thought it was.
It seems the script's DPI settings (passed to inkscape to do batch-exporting) in
render-bitmap.sh are wrong.
It seems that with the current value (DPI=90, the old inkscape's default DPI)
screws up exporting, and the exported small png image is 45x45, NOT 48x48 as it should be.

Bigger icons also suffer from similar scaling issues (eg 120x120 instead of 128x128),
but being bigger it's less noticeable.

Changing that to DPI=96, the current inkscape default dpi,
fixes the issue and the png are created at the right sizes, and shown correctly in my tests.

fixed the theme.conf settings (forgot to comment out options that should have been left commented

I assume that this happens because Inkscape changed since the time the theme (and scripts) were written originally.